+++ src/test/java/uk/gov/hmcts/reform/civil/sampledata/CaseDataBuilder.java
@@ -48,7 +48,6 @@ import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
 
-import static java.math.BigDecimal.TEN;
 import static java.time.LocalDate.now;
 import static uk.gov.hmcts.reform.civil.enums.AllocatedTrack.FAST_CLAIM;
 import static uk.gov.hmcts.reform.civil.enums.CaseState.AWAITING_APPLICANT_INTENTION;
@@ -407,124 +406,6 @@ public class CaseDataBuilder {
         takenOfflineDate = LocalDateTime.now();
         respondent1OrganisationPolicy = null;
         respondentSolicitor1OrganisationDetails = null;
-        return this;
-    }
-
-    public CaseDataBuilder atStateProceedsOfflineUnrepresentedDefendantWithMinimalData() {
-        atStatePaymentSuccessfulWithMinimalData();
-
-        ccdState = PROCEEDS_IN_HERITAGE_SYSTEM;
-        issueDate = CLAIM_ISSUED_DATE;
-        respondent1Represented = NO;
-        takenOfflineDate = LocalDateTime.now();
-        respondent1OrganisationPolicy = null;
-
-        respondentSolicitor1OrganisationDetails = SolicitorOrganisationDetails.builder()
-            .email("testorg@email.com")
-            .organisationName("test org name")
-            .fax("123123123")
-            .dx("test org dx")
-            .phoneNumber("0123456789")
-            .address(AddressBuilder.defaults().build())
-            .build();
-        return this;
-    }
-
-    public CaseDataBuilder atStateClaimDraftWithMinimalData() {
-        courtLocation = CourtLocation.builder()
-            .applicantPreferredCourt("127")
-            .build();
-        applicant1 = PartyBuilder.builder().companyWithMinimalData().build();
-        applicant1LitigationFriendRequired = NO;
-        applicantSolicitor1CheckEmail = CorrectEmail.builder()
-            .email("hmcts.civil@gmail.com")
-            .correct(YES)
-            .build();
-        applicant1OrganisationPolicy = OrganisationPolicy.builder()
-            .organisation(Organisation.builder().organisationID("QWERTY").build())
-            .build();
-        respondent1OrganisationPolicy = OrganisationPolicy.builder()
-            .organisation(Organisation.builder().organisationID("QWERTY").build())
-            .build();
-        respondent1 = PartyBuilder.builder().companyWithMinimalData().build();
-        respondent1Represented = NO;
-        claimType = ClaimType.CLINICAL_NEGLIGENCE;
-        claimValue = ClaimValue.builder()
-            .statementOfValueInPennies(BigDecimal.valueOf(10000000))
-            .build();
-        claimFee = Fee.builder()
-            .calculatedAmountInPence(TEN)
-            .code("fee code")
-            .version("version 1")
-            .build();
-        paymentReference = "some reference";
-        respondentSolicitor1EmailAddress = "respondentsolicitor@example.com";
-        applicantSolicitor1UserDetails = IdamUserDetails.builder().email("applicantsolicitor@example.com").build();
-        applicantSolicitor1ClaimStatementOfTruth = StatementOfTruthBuilder.minimal().build();
-        submittedDate = LocalDateTime.now();
-        return this;
-    }
-
-    public CaseDataBuilder atStatePendingCaseIssuedWithMinimalData() {
-        atStateClaimDraftWithMinimalData();
-        legacyCaseReference = LEGACY_CASE_REFERENCE;
-        allocatedTrack = FAST_CLAIM;
-        ccdState = PENDING_CASE_ISSUED;
-        ccdCaseReference = CASE_ID;
-        return this;
-    }
-
-    public CaseDataBuilder atStatePaymentSuccessfulWithMinimalData() {
-        atStatePendingCaseIssuedWithMinimalData();
-        claimIssuedPaymentDetails = PaymentDetails.builder()
-            .status(SUCCESS)
-            .reference("RC-1604-0739-2145-4711")
-            .build();
-        paymentSuccessfulDate = LocalDateTime.now();
-        claimDetailsNotificationDeadline = LocalDateTime.now().plusDays(1);
-        return this;
-    }
-
-    public CaseDataBuilder atStateAwaitingCaseNotificationWithMinimalData() {
-        atStatePaymentSuccessfulWithMinimalData();
-        ccdState = CASE_ISSUED;
-        issueDate = CLAIM_ISSUED_DATE;
-        claimNotificationDeadline = LocalDateTime.now();
-        return this;
-    }
-
-    public CaseDataBuilder atStateAwaitingCaseDetailsNotificationWithMinimalData() {
-        atStateAwaitingCaseNotificationWithMinimalData();
-        claimNotificationDate = LocalDateTime.now();
-        claimDetailsNotificationDeadline = claimNotificationDate.plusDays(14);
-        ccdState = AWAITING_CASE_DETAILS_NOTIFICATION;
-        return this;
-    }
-
-    public CaseDataBuilder atStateClaimCreatedWithMinimalData() {
-        atStateAwaitingCaseDetailsNotificationWithMinimalData();
-
-        claimDetailsNotificationDate = LocalDateTime.now();
-        claimDismissedDeadline = LocalDateTime.now().plusMonths(6);
-        ccdState = AWAITING_RESPONDENT_ACKNOWLEDGEMENT;
-        respondent1ResponseDeadline = RESPONSE_DEADLINE;
-        return this;
-    }
-
-    public CaseDataBuilder atStateRespondentRespondToClaimWithMinimalData(
-        RespondentResponseType respondentResponseType
-    ) {
-        atStateServiceAcknowledgeWithMinimalData();
-        respondent1ClaimResponseType = respondentResponseType;
-        applicant1ResponseDeadline = APPLICANT_RESPONSE_DEADLINE;
-        respondent1ResponseDate = LocalDateTime.now();
-        ccdState = AWAITING_APPLICANT_INTENTION;
-        return this;
-    }
-
-    public CaseDataBuilder atStateServiceAcknowledgeWithMinimalData() {
-        atStateClaimCreatedWithMinimalData();
-        respondent1ClaimResponseIntentionType = FULL_DEFENCE;
         return this;
     }
