+++ src/test/java/uk/gov/hmcts/reform/civil/sampledata/PartyBuilder.java
@@ -10,7 +10,7 @@ import static uk.gov.hmcts.reform.civil.model.Party.Type.COMPANY;
 import static uk.gov.hmcts.reform.civil.model.Party.Type.INDIVIDUAL;
 import static uk.gov.hmcts.reform.civil.model.Party.Type.ORGANISATION;
 import static uk.gov.hmcts.reform.civil.model.Party.Type.SOLE_TRADER;
-import static uk.gov.hmcts.reform.civil.sampledata.CaseDataEdgeCasesBuilder.MAX_ALLOWED;
+import static uk.gov.hmcts.reform.civil.sampledata.CaseDataMaxEdgeCasesBuilder.MAX_ALLOWED;
 
 public class PartyBuilder {
